[Silver III] Title: 주유소, Time: 348 ms, Memory: 38664 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -17,8 +17,8 @@ public class Main {
             price[i] = Integer.parseInt(part2[i]);
         }
 
-        int min = price[0];
-        int answer = min * distance[0];
+        long min = price[0];
+        long answer = min * distance[0];
 
         for (int i = 1; i < distance.length; i++) {
             if (price[i] < min) {
